+++ src/test/kotlin/sidechain/iroha/consumer/IrohaKeyLoaderTest.kt
@@ -3,6 +3,7 @@ package sidechain.iroha.consumer
 import org.junit.jupiter.api.Assertions.assertNotNull
 import org.junit.jupiter.api.Test
 import sidechain.iroha.IrohaInitialization
+import sidechain.iroha.util.ModelUtil
 
 
 internal class IrohaKeyLoaderTest {
@@ -15,7 +16,7 @@ internal class IrohaKeyLoaderTest {
     @Test
     fun keyLoadFailedTest() {
         IrohaInitialization.loadIrohaLibrary()
-        val keypair = IrohaKeyLoader.loadKeypair("", "")
+        val keypair = ModelUtil.loadKeypair("", "")
 
         assertNotNull(keypair.component2())
     }
